Update Pin_new funcionando as requisições http

+++ pin_new/src/main/java/br/com/pin/repository/RegistroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.pin.controllers.RegistroParada;
 
-public interface RegistroRepository extends JpaRepository<RegistroParada, Long>{
+public interface RegistroRepository extends JpaRepository<RegistroParada, Integer>{
     
 }
